Move permission enum in BoxItem class
