+++ box-content-sdk/src/main/java/com/box/androidsdk/content/models/BoxBookmark.java
@@ -154,62 +154,4 @@ public class BoxBookmark extends BoxItem {
         }
         return mPermissions;
     }
-
-    /**
-     * Enumerates the possible permissions that a user can have on a bookmark.
-     */
-    public enum Permission {
-        /**
-         * The user can rename the bookmark.
-         */
-        CAN_RENAME("can_rename"),
-
-        /**
-         * The user can delete the bookmark.
-         */
-        CAN_DELETE("can_delete"),
-
-        /**
-         * The user can share the bookmark.
-         */
-        CAN_SHARE("can_share"),
-
-        /**
-         * The user can set the access level for shared links to the bookmark.
-         */
-        CAN_SET_SHARE_ACCESS("can_set_share_access"),
-
-        /**
-         * The user can comment on the bookmark.
-         */
-        CAN_COMMENT("can_comment");
-
-        private final String value;
-
-        private Permission(String value) {
-            this.value = value;
-        }
-
-        /**
-         * Returns Permission enum based on a string.
-         *
-         * @param text text to convert to a Permission enum.
-         * @return enum that corresponds to the text.
-         */
-        public static Permission fromString(String text) {
-            if (!TextUtils.isEmpty(text)) {
-                for (Permission a : Permission.values()) {
-                    if (text.equalsIgnoreCase(a.name())) {
-                        return a;
-                    }
-                }
-            }
-            throw new IllegalArgumentException(String.format(Locale.ENGLISH, "No enum with text %s found", text));
-        }
-
-        @Override
-        public String toString() {
-            return this.value;
-        }
-    }
 }
