+++ box-content-sdk/src/main/java/com/box/androidsdk/content/models/BoxFile.java
@@ -236,71 +236,4 @@ public class BoxFile extends BoxItem {
         }
         return mPermissions;
     }
-
-    /**
-     * Enumerates the possible permissions that a user can have on a file.
-     */
-    public enum Permission {
-        /**
-         * The user can download the file.
-         */
-        CAN_DOWNLOAD("can_download"),
-
-        /**
-         * The user can upload new versions of the file.
-         */
-        CAN_UPLOAD("can_upload"),
-
-        /**
-         * The user can rename the file.
-         */
-        CAN_RENAME("can_rename"),
-
-        /**
-         * The user can delete the file.
-         */
-        CAN_DELETE("can_delete"),
-
-        /**
-         * The user can share the file.
-         */
-        CAN_SHARE("can_share"),
-
-        /**
-         * The user can set the access level for shared links to the file.
-         */
-        CAN_SET_SHARE_ACCESS("can_set_share_access"),
-
-        /**
-         * The user can preview the file.
-         */
-        CAN_PREVIEW("can_preview"),
-
-        /**
-         * The user can comment on the file.
-         */
-        CAN_COMMENT("can_comment");
-
-        private final String value;
-
-        private Permission(String value) {
-            this.value = value;
-        }
-
-        public static Permission fromString(String text) {
-            if (!TextUtils.isEmpty(text)) {
-                for (Permission a : Permission.values()) {
-                    if (text.equalsIgnoreCase(a.name())) {
-                        return a;
-                    }
-                }
-            }
-            throw new IllegalArgumentException(String.format(Locale.ENGLISH, "No enum with text %s found", text));
-        }
-
-        @Override
-        public String toString() {
-            return this.value;
-        }
-    }
 }
