+++ box-content-sdk/src/main/java/com/box/androidsdk/content/models/BoxFolder.java
@@ -325,66 +325,4 @@ public class BoxFolder extends BoxItem {
             return this.mValue;
         }
     }
-
-    /**
-     * Enumerates the possible permissions that a user can have on a folder.
-     */
-    public enum Permission {
-        /**
-         * The user can download the folder.
-         */
-        CAN_DOWNLOAD("can_download"),
-
-        /**
-         * The user can upload to the folder.
-         */
-        CAN_UPLOAD("can_upload"),
-
-        /**
-         * The user can rename the folder.
-         */
-        CAN_RENAME("can_rename"),
-
-        /**
-         * The user can delete the folder.
-         */
-        CAN_DELETE("can_delete"),
-
-        /**
-         * The user can share the folder.
-         */
-        CAN_SHARE("can_share"),
-
-        /**
-         * The user can invite collaborators to the folder.
-         */
-        CAN_INVITE_COLLABORATOR("can_invite_collaborator"),
-
-        /**
-         * The user can set the access level for shared links to the folder.
-         */
-        CAN_SET_SHARE_ACCESS("can_set_share_access");
-
-        private final String mValue;
-
-        private Permission(String value) {
-            this.mValue = value;
-        }
-
-        public static Permission fromString(String text) {
-            if (!TextUtils.isEmpty(text)) {
-                for (Permission e : Permission.values()) {
-                    if (text.equalsIgnoreCase(e.toString())) {
-                        return e;
-                    }
-                }
-            }
-            throw new IllegalArgumentException(String.format(Locale.ENGLISH, "No enum with text %s found", text));
-        }
-
-        @Override
-        public String toString() {
-            return this.mValue;
-        }
-    }
 }
